fixing module loading dynamics
