+++ demo/3000-home/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "home",
+  "name": "homeapp",
   "version": "0.1.0",
   "private": true,
   "scripts": {
